refactor:Home.tsx 컴포넌트 분리(#81)
